revert DX to separate link, keep shared code shared
