+++ examples/next-minimal-starter/src/utils/trpc.ts
@@ -1,4 +1,4 @@
-import { httpBatchLink, unstable_streamRequester } from '@trpc/client';
+import { unstable_httpBatchStreamLink } from '@trpc/client';
 import { createTRPCNext } from '@trpc/next';
 import type { AppRouter } from '../pages/api/trpc/[trpc]';
 
@@ -22,9 +22,8 @@ export const trpc = createTRPCNext<AppRouter>({
   config() {
     return {
       links: [
-        httpBatchLink({
+        unstable_httpBatchStreamLink({
           url: getBaseUrl() + '/api/trpc',
-          requester: unstable_streamRequester,
         }),
       ],
     };
