+++ packages/client/src/links/index.ts
@@ -1,7 +1,7 @@
 export * from './types';
 
 export * from './httpBatchLink';
-export * from './httpBatchStreamLink';
+export * from './httpBatchLink/httpBatchStreamLink';
 export * from './httpLink';
 export * from './loggerLink';
 export * from './splitLink';
